Dashboard landing page (#6)
* started side navbar for selecting plants to view

Included a button which when clicked opens a side nav-bar which can be used to select specific plants which will allow the user to view data about the selected plant

* Sidebar added

Sidebar is created and working.

* chart containers showing on sreen

Containers to hold charts are now showing in a flex position so that the layout format stays the same (a grid-like pattern) even when containers are added/removed. Containers currently hold dummy data

* graphs now show on dashboard

Graphs with a header explaining the data alongside the graph with the data points can now be shown with the custom Chart Component